drop old pipeline.
