+++ Sources/TinyPipeline/Duplex.swift
@@ -6,14 +6,12 @@ struct Duplex<Success, Failure> where Failure: Error {
     
     var id = DuplexID()
     
-    #warning("TODO: [Priority: high] delete prefix after removing old future type.")
-    var inbound: (DuplexID, DuplexBoundContext<Success, Failure>) -> TinyCombine.Future<Success, Failure>
+    var inbound: (DuplexID, DuplexBoundContext<Success, Failure>) -> Future<Success, Failure>
     
-    #warning("TODO: [Priority: high] delete prefix after removing old future type.")
-    var outbound: (DuplexID, DuplexBoundContext<Success, Failure>) -> TinyCombine.Future<Success, Failure> = { _, context in
+    var outbound: (DuplexID, DuplexBoundContext<Success, Failure>) -> Future<Success, Failure> = { _, context in
         
         // The default outbound only aggregates the given result.
-        TinyCombine.Future { $0(context.finalResult!) }
+        Future { $0(context.finalResult!) }
         
     }
