+++ Sources/TinyPipeline/DuplexBoundKind.swift
@@ -1,0 +1,7 @@
+// MARK: - DuplexBoundKind
+
+enum DuplexBoundKind {
+    
+    case inbound, outbound
+    
+}
